==> final-notebooks/1a-Key-word_identification.ipynb <==
# Key word identification

In this notebook we identify key words in sentences that are likely to correspond to the Wikipedia article(s) that are most relevant.

import pandas as pd
import spacy
from string import punctuation, digits
import re

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.max_colwidth', 1000)

# Question data
t_data = pd.read_pickle('C:/Users/Fredi/Kodningsprojekt/data-analysis/workproduct-files/cleaned-dataframes/t_dataMaster-duplicatesRemoved.pkl')

# word frequency data
word_frequencies = pd.read_csv("F:/Word frequencies/unigram_freq.csv")

nlp = spacy.load("en_core_web_sm")

#### Functions

def namedEntities(a):
    doc = nlp(a)
    toReturn = []
    
    #If input is title-case, convert to sentence
    if a.istitle():
        a = a.capitalize()
    
    for ent in doc.ents:
        toReturn.append((ent.text, ent.label_))

    return toReturn

def getNouns(a):
    
    #Identify substrings between quotes
    
    quotes1 = re.compile('"[^"]*"')
    quotes2 = re.compile("'[^']*'")
    quoteRanges = []
    toReturn = []
    
    for value in quotes1.finditer(a.replace("'s", "!s")):
        quoteRanges.append(value.span())
        toReturn.append(value.group().replace("!s", "'s")[1:-1])
    for value in quotes2.finditer(a.replace("'s", "!s")):
        quoteRanges.append(value.span())
        toReturn.append(value.group().replace("!s", "'s")[1:-1])
    
    #If input is title-case, convert to sentence
    if a.istitle():
        a = a.capitalize()
    
    doc = nlp(a)
    pos_tag = ['NOUN']
    
    for token in doc:
        if(token.text in nlp.Defaults.stop_words or token.text in punctuation):
            continue
        
        #Exclude words within quotes
        if any(start <= token.idx <= end for (start, end) in quoteRanges):
            continue
        
        if(token.pos_ in pos_tag):
            toReturn.append(token.lemma_)
    
    return toReturn

def getSubjects(a):
    doc = nlp(a.lower())
    toReturn = []
    dep_tag = ['csubj', 'nsubj']
    for token in doc:
        if(token.text in nlp.Defaults.stop_words or token.text in punctuation):
            continue
        if(token.dep_ in dep_tag):
            toReturn.append(token.lemma_)
    
    return toReturn

def getObjects(a):
    doc = nlp(a.lower())
    toReturn = []
    dep_tag = ['iobj', 'obj', 'dobj', 'pobj']
    for token in doc:
        if(token.text in nlp.Defaults.stop_words or token.text in punctuation):
            continue
        if(token.dep_ in dep_tag):
            toReturn.append(token.lemma_)
    
    return toReturn

def frequencyCount(a):
    result = []
    for word in a:
        try:
            result.append([word, word_frequencies.loc[word_frequencies["word"] == word, "count"].iloc[0] ])
        except IndexError:
            # If no count is found, add 0 as value to ensure being sorted first in searchTerms()
            result.append([word, 0])
    return result    
    

# FUNCTION: Identify wikipedia redirect terms to look for (in priority order)
# 1. Check if named entities exist and add in order of priority (WORK_OF_ART, PERSON, FAC, LOC)
# 2. Add nouns in ascending frequency order (rarest first)
# 3. Remove "the" / "a" / "an" / quotes from beginning of named entities

# Takes a question row as input


def searchTerms(a):
    result = []
    
    for ent in a["namedEntities"]:
        if ent[1] == "WORK_OF_ART":
            result.append(ent[0])
    for ent in a["namedEntities"]:
        if ent[1] == "PERSON":
            result.append(ent[0])    
    for ent in a["namedEntities"]:
        if ent[1] == "FAC" or ent[1] == "LOC":
            result.append(ent[0])
    
    # Update to account for objects and subjects
    
    # sort nounsObjectsSubjects by nounFrequencies
    # https://www.kite.com/python/answers/how-to-sort-a-list-of-lists-by-an-index-of-each-inner-list-in-python
    # sorted_list = sorted(nested_list, key=lambda x: x[1])
    wordsSorted = sorted(a["nounsObjectsSubjects"], key=lambda x: x[1])
    for word in wordsSorted:
        result.append(word[0])
        
    # Clean-up, iterate over each value in result list          
    for i in range(len(result)):
        try:
            if result[i][:4].lower() == "the ":
                result[i] = result[i][4:]
            if result[i][:2].lower() == "a ":
                result[i] = result[i][2:]
            if result[i][:3].lower() == "an ":
                result[i] = result[i][3:]
            if result[i][0].lower() == "'" or result[i][0].lower() == '"':
                result[i] = result[i][1:]
            if result[i][-1].lower() == "'" or result[i][-1].lower() == '"':
                result[i] = result[i][:-1]
            if result[i][-2:].lower() == "'s":
                result[i] = result[i][:-2]
        except:
            continue
    return result

#### Identifying key words

questions = t_data

%%time
questions["namedEntities"] = questions["CONS_question"].apply(lambda x: namedEntities(x))

%%time
# Extract namedEntities from answer string
questions["namedEntities_answer"] = questions["CONS_answer"].apply(lambda x: namedEntities(x))

# Combine namedEntities from question and answer into same list
questions["namedEntities"] = questions["namedEntities"] + questions["namedEntities_answer"]

# Drop namedEntities_answer column
questions = questions.drop("namedEntities_answer", axis = 1)

%%time
questions["nouns"] = questions["CONS_question"].apply(lambda x: getNouns(x))

%%time
questions["objects"] = questions["CONS_question"].apply(lambda x: getObjects(x))

%%time
questions["subjects"] = questions["CONS_question"].apply(lambda x: getSubjects(x))

%%time
# Combine nouns, objects and subjects into same list
questions["nounsObjectsSubjects"] = questions["nouns"] + questions["objects"] + questions["subjects"]
# Remove duplicates
questions["nounsObjectsSubjects"] = questions["nounsObjectsSubjects"].apply(lambda x: list(dict.fromkeys(x)))

%%time
# Add frequency count to nounsObjectsSubjects
questions["nounsObjectsSubjects"] = questions["nounsObjectsSubjects"].apply(lambda x: frequencyCount(x))

%%time
questions["searchTerms"] = questions.apply(searchTerms, axis = 1)
# Remove duplicates
questions["searchTerms"] = questions["searchTerms"].apply(lambda x: list(dict.fromkeys(x)))

questions.to_pickle("../workproduct-files/t_dataMaster-keywordsIdentified.pkl")
