==> final-notebooks/4-ScriptRunAutomation.ipynb <==
# Categorization run script

Script to loop through the categorization algorithm for all Wikipedia article and redirect titles. This takes >20h to run. Script is restarted if one run of the algorithm takes >1000 seconds.

# Run Question-to-WikipediaCategory-linking.py
# Infinite loop (or at least until last index is reached)
# Check if result has been updated during x second timelimit
    # If yes, wait and check again after x seconds
    # If no, kill script and start it again

import subprocess
import pandas as pd
import time

sleepTime = 1000
batchRuns = pd.read_csv("../workproduct-files/batchRuns.csv", delimiter=";")
automationRuns = pd.read_csv("C:/Users/Fredi/Google Drive/Diverse/Programmering/CatRunAutomation/automationRuns.csv", delimiter=";")

startingPoint = len(batchRuns)

# First run (assumes neo4j database is started)
subprocess.Popen(['powershell.exe', 'conda activate data-analysis ; cd C:/Users/Fredi/kodningsprojekt/wikipedia-categories/workproduct-notebooks ; python Question-to-WikipediaCategory-linking.py'])

newRow = [batchRuns.iloc[-1, 1], "Manual", time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime())]
automationRuns = automationRuns.append(pd.Series(newRow, index = automationRuns.columns), ignore_index = True)
automationRuns.to_csv("C:/Users/Fredi/Google Drive/Diverse/Programmering/CatRunAutomation/automationRuns.csv", sep=";", index = False)

#This can be improved to check if Question-to-WikipediaCategory-linking is finnished
while True:
    time.sleep(sleepTime)
    batchRuns = pd.read_csv("../workproduct-files/batchRuns.csv", delimiter=";")
    currentPoint = len(batchRuns)
    if currentPoint == startingPoint:
        # Terminate wikipedia-categorization script. (Looks for python.exe using >1GB memory)
        subprocess.Popen(['powershell.exe', 'taskkill /f /fi "IMAGENAME eq python.exe" /fi "MEMUSAGE gt 1000000"'])
        
        time.sleep(10)
        
        # Restart script after waiting 10 seconds
        subprocess.Popen(['powershell.exe', 'conda activate data-analysis ; cd C:/Users/Fredi/kodningsprojekt/wikipedia-categories/workproduct-notebooks ; python Question-to-WikipediaCategory-linking.py'])
        
        newRow = [batchRuns.iloc[-1, 1], "Auto", time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime())]
        automationRuns = automationRuns.append(pd.Series(newRow, index = automationRuns.columns), ignore_index = True)
        automationRuns.to_csv("C:/Users/Fredi/Google Drive/Diverse/Programmering/CatRunAutomation/automationRuns.csv", sep=";", index = False)
        
        print("restart at " + str(time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime())) + " and index " + str(batchRuns.iloc[-1, 1]) )
    else:
        startingPoint = currentPoint
