==> final-notebooks/6-td-MasterData.ipynb <==
# Data export a

Creation of data output for next stage of project.

import pandas as pd
import numpy as np
import seaborn as sns
import time
import glob
import re

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.max_colwidth', 1000)

Run = "../workproduct-files/batchRuns/"


RunMeta = pd.read_csv("../workproduct-files/batchRuns.csv", delimiter=";")


t_data = pd.read_pickle("../workproduct-files/t_dataMaster-keywordsIdentified.pkl")

RunMeta

Run_time = RunMeta.iloc[:,-1].sum(skipna = True)

#Total run time in days
Run_time/3600/24

#Days at average speed per row
Run_time/RunMeta.iloc[-1,1]*65000/3600/24

Run_files = glob.glob(Run + "*.pkl")


%%time
RunDF = pd.DataFrame()
for filename in Run_files:
    RunDF = RunDF.append(pd.read_pickle(filename))
RunDF = RunDF.sort_index()

Run_wikisearchFailed = len(RunDF[RunDF.wikipediaSearchSuccessful == False])
Run_wikisearchFailed

def successIndex(a):
    try:
        result = a.index("SUCCESS")
    except ValueError:
        result = -1
    
    return result

RunDF["searchTermUsed_idx"] = RunDF["findQuestionCategories_meta"].apply(lambda x: successIndex(x))

%%time
for row in RunDF.index:
    if RunDF.loc[row, "searchTermUsed_idx"] != -1:
        RunDF.loc[row, "searchTermUsed"] = RunDF.loc[row, "searchTerms"][RunDF.loc[row, "searchTermUsed_idx"]]
    else:
        RunDF.loc[row, "searchTermUsed"] = None
    

RunDF

#Combine dataframes leaving out searchTerms from RunDF (duplicate in result)
t_dataRun = pd.concat([t_data, RunDF.iloc[:,1:]], axis = 1, join = "inner")

# Suboptimal solution, only for preanalyses. If excludes iloc operation on rows with value None
t_dataRun["WikipediaHighestCat"] = t_dataRun["categoryPath"].apply(lambda x: x.iloc[-2, 0] if (type(x) is not type(None)) else None)

t_dataRun["categoryPathList"] = t_dataRun["categoryPath"].apply(lambda x: list(reversed(x.iloc[:,0])) if (type(x) is not type(None)) else [])

t_dataRun["categoryPathList"]

# For each category level in max level count
# Add column with level data
for lvl in range(max(t_dataRun["categoryPathList"].apply(lambda x: len(x)))):
    t_dataRun["WikiCat_" + str(lvl)] = t_dataRun["categoryPathList"].apply(lambda x: x[lvl] if len(x) > lvl else None)

t_dataRun.columns

# CATEGORIZATION RULES

## Wikipedia path final category

# Wikipedia category level 1 mapping
# See spreadsheet https://docs.google.com/spreadsheets/d/1c4g4MujTgaqTn5wG5I-ARY06g7rlnHst0whojsFbps8/edit?usp=sharing
MAPPING_wiki_to_trivia = {'History':'History and society' ,'Reference':'None' ,'Law':'History and society' ,'Religion':'History and society' ,'Education':'History and society' ,'Philosophy':'Science and nature' ,'Politics':'History and society' ,'Science_and_technology':'Science and nature' ,'Sports':'Sports and hobbies' ,'Language':'Art and literature' ,'Nature':'Science and nature' ,'Geography':'Geography and places' ,'Arts':'Art and literature' ,'Society':'History and society' ,'Culture':'History and society' ,'Health':'Science and nature' ,'Mathematics':'Science and nature' ,'Universe':'Science and nature' ,'Events':'History and society' ,'Humanities':'Art and literature' ,'Life':'Science and nature' ,'Entertainment':'Entertainment and games' }


t_dataRun["WikiPath_Category"] = t_dataRun["WikiCat_1"].replace(MAPPING_wiki_to_trivia)

## Wikipedia path exceptions

# https://docs.google.com/spreadsheets/d/1c4g4MujTgaqTn5wG5I-ARY06g7rlnHst0whojsFbps8/edit#gid=2055637170
# Mapping list
exceptionCategories = [('3', 'Literature', 'Art and literature'),('2', 'Food_and_drink', 'Sports and hobbies'),('5', 'Plants', 'Science and nature'),('3', 'Ethnobiology', 'Science and nature'),('2', 'Artificial_objects', 'Science and nature'),('4', 'Diseases_and_disorders', 'Science and nature'),('2', 'Performing_arts', 'Art and literature'),('2', 'Gaming', 'Entertainment and games'),('4', 'Anthropology_of_religion', 'History and society'),('3', 'Leisure', 'Sports and hobbies'),('4', 'Physical_geography', 'Geography and places')]

# Take dataframe row as input
# If several exceptions apply, algorithm returns the one listed last in exceptionCategories
def findExceptionWikiCatRows(a):
    
    result = ""
    
    for exep in exceptionCategories:
        if a["WikiCat_" + str(exep[0])] == exep[1]:
            result = exep[2]
    
    return result
    

row = t_dataRun.iloc[4277]
findExceptionWikiCatRows(row)

%%time
t_dataRun["WikiPath_Exceptions"] = t_dataRun.apply(lambda x: findExceptionWikiCatRows(x), axis = 1)

%%time
# If exception exists, use that, else use WikiPath_Category
t_dataRun["WikiPath_withExceptions"] = t_dataRun.apply(lambda x: x["WikiPath_Exceptions"] if x["WikiPath_Exceptions"] != "" else x["WikiPath_Category"], axis = 1)

# t_dataRun.drop("WikiPath_Exceptions", axis = 1, inplace = True)

#crosstab_percentCols = pd.crosstab(t_dataRun.WikiPath_Category, t_dataRun.WikiPath_withExceptions, margins=True, normalize = "columns").round(3)*100
crosstab_percentCols = pd.crosstab(t_dataRun.WikiPath_Category, t_dataRun.WikiPath_withExceptions)
crosstab_percentCols.style.background_gradient(cmap='Blues', axis = 0)

t_dataRun["WikiPath_Exceptions"].groupby(t_dataRun["WikiPath_Exceptions"]).count()

## Key word exceptions

# Except specific search term / key word


### Category is History if Q / A contains number within range

# 1970 as upper bound seems better. Many questions regarding 1970's are better categorized elsewhere.
historyLowerBound = 1000
historyUpperBound = 1970

# Takes list of numbers as input
def withinHistoryRange(a):
    for x in a:
        try:
            if int(x) > historyLowerBound and int(x) < historyUpperBound:
                return True
        except:
            return "ERROR - Input was probably not a number."
    return False

%%time
t_dataRun["hasHistoryNumber"] = (t_dataRun["CONS_question"] + t_dataRun["CONS_answer"]).apply(lambda x: withinHistoryRange(re.findall(r'\d+', x)))

# NOTE: Categorization currently not impacted by presence of historical date

### Search term exceptions

# Search term mapping: https://docs.google.com/spreadsheets/d/1c4g4MujTgaqTn5wG5I-ARY06g7rlnHst0whojsFbps8/edit#gid=2055637170

termMappingList = {'capital':'Geography and places' ,'state':'Geography and places' ,'sport':'Sports and leisure' }

# Take search term list as input
# If several exceptions apply, algorithm returns the one corresponding to search term listed first
def findExceptionSearchTerms(a):
           
    for term in a:
        if term in termMappingList:
            return (termMappingList[term], term)
    
    return None
    

%%time
t_dataRun["searchTermCatExceptions"] = t_dataRun["searchTerms"].apply(lambda x: findExceptionSearchTerms(x))

t_dataRun["WikiPath_withExceptions"].unique()

### Combine original category and wikicategory fields

t_dataRun["CONS_category"].unique()

# Change category names to original category data
MAPPING_old_cat_to_new = {'Entertainment': 'Entertainment and games', 'Sports and leisure':'Sports and hobbies'}
t_dataRun["CONS_category"] = t_dataRun["CONS_category"].replace(MAPPING_old_cat_to_new)

t_dataRun["CONS_category"].unique()

t_dataRun.groupby('CONS_category').count().iloc[:,0]

t_dataRun.pivot_table(values='CONS_id', index=['Source', 'CONS_category'], 
                         aggfunc=pd.Series.nunique)

# Set Master_category to CONS_category if it has a value, otherwise to WikiPath_withExceptions
t_dataRun["Master_category"] = t_dataRun["CONS_category"]

t_dataRun.loc[t_dataRun.CONS_category == "Uncategorized", "Master_category"] = t_dataRun.loc[t_dataRun.CONS_category == "Uncategorized", "WikiPath_withExceptions"]

t_dataRun["Master_category"].unique()

# Add column Language with value of en for english
t_dataRun["Language"] = "en"

# Add column Row_id that equals index and drop columns CONS_id, ORIG_id, Duplicate_removed and Source
t_dataRun.reset_index(inplace=True)

t_dataRun = t_dataRun.rename(columns = {'index':'Row_id'})

t_dataRun.drop(columns=['CONS_id', 'ORIG_id', 'Duplicate_removed', 'Source'], inplace=True)

### Identify questions to use in v1.0

Criteria:
- Has a "Master_category"
- "CONS_type-formulation" is 'Question' or 'Statement - open'
- "CONS_type-multipleChoice" is False

t_dataRun["CONS_type-formulation"].unique()

t_dataRun["CONS_type-multipleChoice"].unique()

acceptedCategories = ["Art and literature", "Entertainment and games", "Geography and places", "History and society", "Science and nature", "Sports and hobbies"]
#Exclude statement-boolean (true/false questions)
acceptedCONS_typeFormulation = ['Question', 'Statement - open']
#Include both multiple choice and others
acceptedCONS_typeMultipleChoice = [False, True]

def isPrimarySet(a):
    try:
        out = a["Master_category"] in acceptedCategories and a["CONS_type-formulation"] in acceptedCONS_typeFormulation and a["CONS_type-multipleChoice"] in acceptedCONS_typeMultipleChoice
    except:
        out = "ERROR"
    
    return out

%%time
#t_dataRun["primarySet"] = t_dataRun.apply(lambda x: x["Master_category"] in acceptedCategories and x["CONS_type-formulation"] in acceptedCONS_typeFormulation and x["CONS_type-multipleChoice"] in acceptedCONS_typeMultipleChoice)
t_dataRun["primarySet"] = t_dataRun.apply(lambda x: isPrimarySet(x), axis = 1)

t_dataRun.pivot_table(values='Row_id', index=['primarySet'], 
                         aggfunc=pd.Series.nunique)

### Master data export

# Export as pickle, csv and json to new folder (final-data)

t_dataRun.to_pickle("../final-files/tdMasterData_2020-12-16.pkl")

t_dataRun.to_json(path_or_buf="../final-files/tdMasterData_2020-12-16.json", orient="records")

t_dataRun.to_csv(path_or_buf="../final-files/tdMasterData_2020-12-16.csv", index=False)
