==> final-notebooks/7-td-static-data-export.ipynb <==
# Data export b

Creation of data output for next stage of project.

import pandas as pd
import time

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.max_colwidth', 1000)

%%time
t_data = pd.read_pickle("../workproduct-files/t_dataMaster-keywordsIdentified.pkl")
master_data = pd.read_pickle("../final-files/tdMasterData_2020-12-16.pkl")

rows = t_data.loc[t_data["Source"]=="rtg"].index.tolist()

out = master_data.iloc[rows]

out = out.loc[out["primarySet"] == True]

out = out.loc[:, ['CONS_question', 'CONS_answer', 'Master_category']]

out.groupby('Master_category').count()

sample = out.sample(frac=1).groupby('Master_category').head(150)

sample

# Change column names
sample.rename(columns={'CONS_question': 'question', 'CONS_answer': 'answer', 'Master_category': 'category'}, inplace=True)

# Remove potential formatting errors
sample["question"] = sample["question"].apply(lambda x: x.replace("â€™", "'"))

# Save as JSON
sample.to_json("../workproduct-files/td-static-data-export_v1.json", orient="records")
